Modification Paraglider Poids en decimal

+++ ParaglidingProject.SL.Core/ParagliderModel.NS/TransfertObjects/ParagliderModelDto.cs
@@ -8,8 +8,8 @@ namespace ParaglidingProject.SL.Core.ParagliderModel.NS.TransfertObjects
   {
     public int ID { get; set; }
     public string Size { get; set; }
-    public int MaxWeightPilot { get; set; }//Rien dans Models configuration
-    public int MinWeightPilot { get; set; } 
+    public decimal MaxWeightPilot { get; set; }//Rien dans Models configuration
+    public decimal MinWeightPilot { get; set; } 
     public string ApprovalNumber { get; set; }
     public DateTime ApprovalDate { get; set; }
     public bool IsActive { get; set; }
